+++ NEWS.md
@@ -3,6 +3,7 @@
 * Disallow additional properties in item requests
 * Disallow additional properties in instance requests
 * Disallow additional properties in loan type requests
+* Disallow additional properties in material type requests
 * Upgrade to RAML Module Builder 11.0.0
 * Upgrade to cql2pgjson 1.1.0